+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library("testthat")
-library("enrichwith")
-
-test_check("enrichwith")